BUGFIX fix the generator for prop_mod_inv
It could very easily get stuck when testing more than 100 times. For 100 times
it was less prone to fail, but maybe 1 out of 10 runs in travis failed due to
this problem.

The new procedure to generate coprimes generates first two non-overlapping lists
of prime factors, so we know for sure gcd is 1. This has the advantage that we
will be testing big numbers as well, with the previous generator it was very
unlikely that we could find a pair of big coprimes.